# DeBroglieProject/docs/swagger.yaml
basePath: /api
definitions:
...
  ds.User:
    properties:
      email:
        type: string
      is_moderator:
        type: boolean
      name:
        type: string
      uuid:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчета длины волны де Бройля частиц
  title: De Broglie Project API
  version: "1.0"
paths:
...
  /profile/me:
    get:
      description: Возвращает информацию о авторизованном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            $ref: '#/definitions/ds.User'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о текущем пользователе
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Обновляет информацию о авторизованном пользователе
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.User'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная информация о пользователе
          schema:
            $ref: '#/definitions/ds.User'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление информации о текущем пользователе
      tags:
      - Profile
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"